+++ lib/vanagon/platform/defaults/amazon-7-aarch64.rb
@@ -1,0 +1,10 @@
+platform "amazon-7-aarch64" do |plat|
+  plat.servicedir "/usr/lib/systemd/system"
+  plat.defaultdir "/etc/sysconfig"
+  plat.servicetype "systemd"
+
+  packages = %w(autoconf automake createrepo gcc gcc-c++ rsync cmake make rpm-libs rpm-build libarchive)
+  plat.provision_with("yum install -y --nogpgcheck  #{packages.join(' ')}")
+  plat.install_build_dependencies_with "yum install --assumeyes"
+  plat.vmpooler_template "amazon-7-arm64"
+end